Enhance LLM support in main module.
- Initialize LLM client.
- Replace GenerateOpenAIResponse with GenerateLLMResponse.
- Move structs to type.go.
